requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
basicsr==1.4.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
cmake==3.25.2
contourpy==1.1.0
controlnet-aux==0.0.6
cycler==0.11.0
decorator==5.1.1
diffusers==0.18.2
dill==0.3.6
einops==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
googleapis-common-protos==1.56.4
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==4.11.3
importlib-resources==6.0.0
iniconfig==2.0.0
intel-openmp==2023.1.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.3
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.2
mkl==2023.0.0
mkl-devel==2023.0.0
mkl-include==2023.0.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numpy==1.23.1
omegaconf==2.3.0
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pluggy==1.2.0
protobuf==3.20
psutil==5.9.5
pyahocorasick==2.0.0
pyarrow==12.0.1
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydeck==0.8.1b0
Pygments==2.15.1
pylint==2.17.3
Pympler==1.0.1
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
pytorch-lightning==2.0.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.3.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
streamlit==1.24.1
sympy==1.12
tbb==2021.9.0
tenacity==8.2.2
tifffile==2023.7.10
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torchmetrics==1.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.23.1
validators==0.20.0
watchdog==3.0.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
